% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eb_times.r
\name{get_eb_times}
\alias{get_eb_times}
\title{Retrieve event boundary times from a segmag object}
\usage{
get_eb_times(segmag, cutoff)
}
\arguments{
\item{segmag}{object of class \code{\link{segmag}}}

\item{cutoff}{numeric value determining the critical cutoff in segmentation
magnitude}
}
\value{
numeric vector with event boundary times
}
\description{
Returns the times of event boundaries from a segmag object. Event boundaries
are defined as the center of local maxima in segmentation magnitude that are
higher than a critical cutoff value. The critical cutoff value should be
determined with the \code{\link{bootstrap_critical_cutoffs}} function.
}
\examples{
#see ?segmag for an example
}
\seealso{
\code{\link{get_eb_times_segmag_diff}}, \code{\link{bootstrap_critical_cutoffs}}
}

